[#58] FEAT: 선택한 옵션을 체크하는 메서드 추가
